+++ CHANGELOG.md
@@ -54,6 +54,9 @@
 * Incorrect multi-column character spacer insertion when reflowing
   text.
 * Compilation errors in 32-bit builds.
+* Mouse cursor style in top and left margins.
+* Selection is now **updated** when the cursor moves outside the grid
+  (https://codeberg.org/dnkl/foot/issues/70).
 
 
 ### Security